# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Users Module API
  description: |
    REST API for the users-module. Manages user accounts, authentication,
    authorization, application tenancy, and audit logging.
  version: 0.1.0
  contact:
    email: zane@liquid-labs.com
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT obtained via any supported auth method (password, email-code,
        Google OIDC, Microsoft OIDC). Pass as `Authorization: Bearer <token>`.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          example: bad_request
        message:
          type: string
          description: Human-readable error description.
          example: "invalid JSON body"
...
    UserSummary:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        is_admin:
          type: boolean
        entity:
          $ref: "#/components/schemas/EntityInfo"
    EntityInfo:
      type: object
      properties:
        kind:
          type: string
          enum: [natural_person, corporation, service_account]
        display_name:
          type: string
        given_name:
          type: string
        family_name:
          type: string
        legal_name:
          type: string
        jurisdiction:
          type: string
        label:
          type: string
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          properties:
            default_app:
              $ref: "#/components/schemas/AppSummary"
...
    AdminUpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        given_name:
          type: string
        family_name:
          type: string
        is_admin:
          type: boolean
...
    AppSummary:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    AppDetail:
      allOf:
        - $ref: "#/components/schemas/AppSummary"
        - type: object
          properties:
            created_at:
              type: string
              format: date-time
            archived_at:
              type: string
              format: date-time
              nullable: true
    AppCreateRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
    AppUpdateRequest:
      type: object
      properties:
        name:
          type: string
...
  parameters:
    UserUUID:
      name: user_uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AppUUID:
      name: app_uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Caller lacks permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /v1/users/{user_uuid}:
    get:
      summary: Get a user by UUID (admin)
      tags: [Users]
      operationId: getUser
      parameters:
        - $ref: "#/components/parameters/UserUUID"
      responses:
        "200":
          description: User detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a user (admin)
      tags: [Users]
      operationId: updateUser
      parameters:
        - $ref: "#/components/parameters/UserUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Archive a user (admin, soft-delete)
      tags: [Users]
      operationId: archiveUser
      parameters:
        - $ref: "#/components/parameters/UserUUID"
      responses:
        "204":
          description: User archived.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/apps:
    get:
      summary: List applications (admin)
      tags: [Apps]
      operationId: listApps
      responses:
        "200":
          description: List of applications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create an application (admin)
      tags: [Apps]
      operationId: createApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppCreateRequest"
      responses:
        "201":
          description: Application created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/apps/{app_uuid}:
    get:
      summary: Get application detail (admin)
      tags: [Apps]
      operationId: getApp
      parameters:
        - $ref: "#/components/parameters/AppUUID"
      responses:
        "200":
          description: Application detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update an application (admin)
      tags: [Apps]
      operationId: updateApp
      parameters:
        - $ref: "#/components/parameters/AppUUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppUpdateRequest"
      responses:
        "200":
          description: Updated application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Archive an application (admin, soft-delete)
      tags: [Apps]
      operationId: archiveApp
      parameters:
        - $ref: "#/components/parameters/AppUUID"
      responses:
        "204":
          description: Application archived.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...